Method for models to be removed from the world
A model now have a `status` which is (currently) either Model.KEEP or
model.REMOVE. When it is to be removed the iterating loop of the Scene
checks the status of the model and removes it if needed.

The `View` also removes models which are not 'active' anymore from their
lists of to-be-rendered actors.