Pause/unpause with same key works
Some infrastructure changes.

First, a Scene always updates/renders/on_events, even when not active.
It's the subclass' duty to check for `self.is_active()` if it wants to
not render for example or whatever.

The Game scene can't be paused anymore, it can only be
activate/inactive.

The activation mechanism now works with a private `__active_toggle`
variable. It determines the toggle of the active state upon next
`update()`. So, each game frame, you are still `is_active()` even if you
`active(False)`, until the next `update()`.

I think that's more or less all of it.